+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 BUG FIXES:
 
 * resource/detector: Fixed incorrect documentation for Slack notifications. Thanks [gpetrousov](https://github.com/gpetrousov). [#25](https://github.com/terraform-providers/terraform-provider-signalfx/issues/25)
-
+* resource/detector: Fixed invalid field for OpsGenie notifications. [#16](https://github.com/terraform-providers/terraform-provider-signalfx/issues/16)
 
 ## 4.1.0 (July 17, 2019)